refactor: clippy clean — the parent-mismatch check compared the same pair twice, and every warning fixed
- fsck: 'actual != claimed_parent && claimed_parent != actual' was the
  same test written twice (clippy::eq_op, an error); one comparison.
- fs: read_inode_data filled holes with iter::repeat_n, stable only in
  1.82 against a 1.75 MSRV; Vec::resize does the same.
- csum: the software crc32c is compiled for no_std and for the host test
  that holds it to the accelerated one, so it is no longer dead under std.
- format: the csum-seed branches were identical; a type alias for the
  placement cache; iterate the pointer array instead of indexing it.
- read: BlockReader's Result<(), ()> is deliberate for firmware and says
  why; a same-type cast and a struct-update initialiser.
- htree: the half-MD4 rounds use rotate_left.
- device, extent: same-type cast, arrays where vec! was used.

+++ src/device.rs
@@ -173,7 +173,7 @@ fn detect_sector_size(file: &std::fs::File) -> u32 {
         {
             if let Ok(size) = rustix::fs::ioctl_blksszget(file) {
                 if size > 0 {
-                    return size as u32;
+                    return size;
                 }
             }
         }
